modified stepper element to use NEMA specific sizes
