+++ src/index.ts
@@ -50,4 +50,3 @@ export { DipSwitch8Element } from './dip-switch-8-element';
 export { StepperMotorElement } from './stepper-motor-element';
 export { HX711Element } from './hx711-element';
 export { KS2EMDC5Element } from './ks2e-m-dc5-element';
-export { NemaStepperElement } from './nema-stepper-element';
